--- a/src/python/pants/backend/go/util_rules/third_party_pkg.py
+++ b/src/python/pants/backend/go/util_rules/third_party_pkg.py
@@ -595,15 +595,11 @@ async def download_and_analyze_third_party_packages(
         )
     )
 
-    go_mod_digest = await merge_digests(
-        MergeDigests([request.go_mod_digest, module_analysis.go_mods_digest])
-    )
-
     analyzed_modules = await concurrently(
         analyze_go_third_party_module(
             AnalyzeThirdPartyModuleRequest(
                 go_mod_address=request.go_mod_address,
-                go_mod_digest=go_mod_digest,
+                go_mod_digest=request.go_mod_digest,
                 go_mod_path=request.go_mod_path,
                 import_path=mod.name,
                 name=mod.name,